+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changes to Calva.
 - [Support for custom project/workflow commands](https://github.com/BetterThanTomorrow/calva/issues/281)
 
 ## [Unreleased]
+- [Better inline evaluation error reports with file context](https://github.com/BetterThanTomorrow/calva/issues/329)
 
 ## [2.0.39] - 20.09.2019
 - [Revert disconnecting and jacking out on closing of REPL window](https://github.com/BetterThanTomorrow/calva/issues/326)