+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to Calva.
 
 ## [Unreleased]
 - [Escape HTML in stdout and stderr in REPL window](https://github.com/BetterThanTomorrow/calva/issues/321)
+- [Add hover to inline result display, containing the full results](https://github.com/BetterThanTomorrow/calva/pull/336)
 
 ## [2.0.39] - 20.09.2019
 - [Revert disconnecting and jacking out on closing of REPL window](https://github.com/BetterThanTomorrow/calva/issues/326)